Update Calculator(Windows Form App)
This can be termed as a calculator which can be used in windows form application.  The language used in this code is c#.This calculator is very modern in its approach as it can calculate with both buttons and text boxes , like that of in android and in desktops used by us.